==> code/02_Data Cleaning.ipynb <==
## Data Cleaning


import pandas as pd
import re

This imports  the libraries that we need. 

hurricane_harvey_pre = pd.read_csv('../data/hurricane_harvey_pre.csv',delimiter=";;")
hurricane_irma_pre = pd.read_csv('../data/hurricane_irma_pre.csv',delimiter=";;")
hurricane_maria_pre = pd.read_csv('../data/hurricane_maria_pre.csv',delimiter=";;")
north_bay_2014_earthquake_pre = pd.read_csv('../data/north_bay_2014_earthquake_pre.csv',delimiter=";;")
prague_2011_earthquake_pre = pd.read_csv('../data/prague_2011_earthquake_pre.csv',delimiter=";;")
wildfire_campfire_pre = pd.read_csv('../data/wildfire_campfire_pre.csv',delimiter=";;")
wildfire_carrfire_pre = pd.read_csv('../data/wildfire_carrfire_pre.csv',delimiter=";;")
wildfire_woolseyfire_pre = pd.read_csv('../data/wildfire_woolseyfire_pre.csv',delimiter=";;")

This imports the pre-disaster data that we scraped using GetOldTweets3 into DataFrames. 

combined_pre = pd.concat([hurricane_harvey_pre, hurricane_irma_pre,hurricane_maria_pre,north_bay_2014_earthquake_pre,prague_2011_earthquake_pre,wildfire_campfire_pre,wildfire_carrfire_pre,wildfire_woolseyfire_pre], ignore_index=True)

This creates a combined dataframe of all of the pre-disaster Dataframes that we previously imported. 

combined_pre['text'] = combined_pre.text.map(lambda x: re.sub('http[s]?:\/\/[^\s]*', ' ', x))

combined_pre['text'] = combined_pre.text.map(lambda x: re.sub('[^A-z]', ' ', x))

This removes all URLs and nonalphabetical characters from our pre-disaster twitter texts.

hurricane_harvey_post = pd.read_csv('../data/hurricane_harvey_post.csv',delimiter=";;")
hurricane_irma_post = pd.read_csv('../data/hurricane_irma_post.csv',delimiter=";;")
hurricane_maria_post = pd.read_csv('../data/hurricane_maria_post.csv',delimiter=";;")
north_bay_2014_earthquake_post = pd.read_csv('../data/north_bay_2014_earthquake_post.csv',delimiter=";;")
prague_2011_earthquake_post = pd.read_csv('../data/prague_2011_earthquake_post.csv',delimiter=";;")
wildfire_campfire_post = pd.read_csv('../data/wildfire_campfire_post.csv',delimiter=";;")
wildfire_carrfire_post = pd.read_csv('../data/wildfire_carrfire_post.csv',delimiter=";;")
wildfire_woolseyfire_post = pd.read_csv('../data/wildfire_woolseyfire_post.csv',delimiter=";;")

This imports the post-disaster data that we scraped using GetOldTweets3 into DataFrames. 

combined_post = pd.concat([hurricane_harvey_post, hurricane_irma_post,hurricane_maria_post,north_bay_2014_earthquake_post,prague_2011_earthquake_post,wildfire_campfire_post,wildfire_carrfire_post,wildfire_woolseyfire_post], ignore_index=True)

This creates a combined dataframe of all of the post-disaster Dataframes that we previously imported. 

combined_post['text'] = combined_post.text.map(lambda x: re.sub('http[s]?:\/\/[^\s]*', ' ', x))

combined_post['text'] = combined_post.text.map(lambda x: re.sub('[^A-z]', ' ', x))

This removes all URLs and nonalphabetical characters from our post-disaster twitter texts. 

combined_pre.to_csv('../assets/combined_pre.csv')

combined_post.to_csv('../assets/combined_post.csv')

This exports our pre and post disaster combined DataFrames to csv files.
