==> code/03_Data_Classifying.ipynb <==
import pandas as pd
import nltk

This imports the libraries we need. 

## Combined_pre Preparation

combined_pre = pd.read_csv('../assets/combined_pre.csv', index_col=0)
combined_pre = combined_pre[['text']]
combined_pre['text'] = combined_pre['text'].apply(str.lower)
combined_pre.drop_duplicates(subset='text', inplace=True)
combined_pre['disaster'] = 0
combined_pre.reset_index()
combined_pre.head()

This classifies all of the predisaster data as being not disaster related and also drops all duplicates. 

## Combined_post Preparation

combined_post = pd.read_csv('../assets/combined_post.csv', index_col=0)
combined_post = combined_post[['text']]
combined_post['text'] = combined_post['text'].apply(str.lower)
combined_post.drop_duplicates(subset='text', inplace=True)
combined_post.head()

disaster_keywords = []
with open('../assets/disaster_keywords.txt') as f:
    for phrase in f:
        disaster_keywords.append(phrase.strip())
disaster_keywords[:5]

def disaster_related(tweet): 
    words = set(nltk.word_tokenize(tweet))
    if len(words.intersection(set(disaster_keywords))) > 0: 
        return 1
    return 0

This is a function that will classify the postdisaster data as disaster related or not. 

combined_post['disaster'] = combined_post['text'].apply(disaster_related)

This classifies the postdisaster data as being disaster related or not. 

combined_post.head()

print(combined_pre.shape)
print(combined_post.shape)

combined_post['disaster'].value_counts(normalize=True)

## Saving the files to csvs

combined_pre.to_csv('../assets/combined_pre_classed.csv')
combined_post.to_csv('../assets/combined_post_classed.csv')


This saves the predisaster and postdisaster dataframes to csv format.
