==> code/04b_Semantic Analysis_LDA.ipynb <==
We will import the libraries needed throughout this notebook.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import pickle
import nltk
from nltk.corpus import stopwords
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation

We opened the pickled customized stopwords created from the previous notebook. 

with open('./assets/stopwords.pkl','rb') as f:
    stopwords = pickle.load(f)

This is combined dataframe containing both the pre-disaster and post-disaster tweets.

combined_df = pd.read_csv('../project_4/assets/combined_df.csv')

combined_df.shape

Once again we will implement the TFIDF tool to have the words recored as integer counts in order to be used on a the Logistic Regression model. We will also split our data into 'X' and 'y'. The X is an object that consists of the predictor column, in this case the text (tweets). The y will contain the binary classifier column, whether or not the tweets belongs to a disaster class or non-disaster class.
We will randomly split the data into training and test sets. This is done so we can train our model on the training set and then evaluate the performance of the model on unseen new data (the validation set).

y = combined_df['disaster']

# Set X as text column.
X = combined_df['text']

X_train, X_test, y_train, y_test = train_test_split(X,
                                                   y, 
                                                    test_size=.30,
                                                   random_state=42,
                                                   stratify=y)

tfidf = TfidfVectorizer(stop_words = stopwords, 
                        max_df=0.95, 
                        min_df=5, max_features=10_000)

X_train_tfidf = tfidf.fit_transform(X_train)
X_test_tfidf = tfidf.transform(X_test)

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression()
model = lr.fit(X_train_tfidf, y_train)

print(f'LogReg Training score: {model.score(X_train_tfidf, y_train)}')
print(f'LogReg Testing score: {model.score(X_test_tfidf, y_test)}')

-------

## LDA:

Latent Dirichlet Allocation (LDA) is a form of feature extraction. We will apply LDA on a corpos of documents (tweets) and extract from it additive models of the topic structure of the corpus (collection of documents). The outcome will be a different topics and each of those topics are represented as a list of terms.

X_train_df_tfidf = pd.SparseDataFrame(X_train_tfidf,
                             columns=tfidf.get_feature_names())

X_train_df_tfidf.shape

X_train_df_tfidf.head()

X_train_df_tfidf.fillna(0, inplace=True)

X_test_df_tfidf = pd.SparseDataFrame(X_test_tfidf,
                                    columns = tfidf.get_feature_names())

X_test_df_tfidf.fillna(0, inplace=True)
print(X_test_df_tfidf.shape)

LDA is a method of discovering topics from sentences. This is applicable to our project so we can better understand and identify the different topics of the tweets in our dataset. Due to our constraints of time and lack of in depth knowledge on LDA, we did not experiment with all the parameters of LDA. 



n_features = 1000
n_components = 10
n_top_words = 20


def print_top_words(model, feature_names, n_top_words):
    for topic_idx, topic in enumerate(model.components_):
        message = "Topic #%d: " % topic_idx
        message += " ".join([feature_names[i]
                             for i in topic.argsort()[:-n_top_words - 1:-1]])
        print(message)
    print()


# Use tf (raw term count) features for LDA.
print("Extracting tf features for LDA...")
tf_vectorizer = TfidfVectorizer(max_df=0.95, min_df=5,
                                max_features=n_features,
                                stop_words=stopwords)

X_train_tf = tf_vectorizer.fit_transform(X_train)
X_test_tf = tf_vectorizer.transform(X_test)


lda = LatentDirichletAllocation(n_components=n_components, max_iter=5,
                                learning_method='online',
                                random_state=42)

lda_train = lda.fit_transform(X_train_tf)
lda_test = lda.transform(X_test_tf)


print("\nTopics in LDA model:")
tf_feature_names = tf_vectorizer.get_feature_names()
print_top_words(lda, tf_feature_names, n_top_words)

According to this output, most of the topics are related to the hurricanes. Almost all of the 10 topics seem to be somehow related to the hurricane natural disaster, except topic 3. Topic 3 seems to relate more with the mass shooting that took place in the city of Thousand Oaks. 

Although LDA seems to be a popular method when it comes to textual analysis/language processing. LDA is primarily utilized to find latent (hidden) topics in the documents. Unfortunately it seems that LDA, where each word has a topic label, may not work well with Twitter as Twitter messages (tweets) are short and a single tweet is more likely to talk about one topic.

feature_loadings = pd.DataFrame(lda.components_, 
                                columns = tf_vectorizer.get_feature_names(),
                                index = [f'topic_{x}' for x in range(lda.components_.shape[0])]).T

feature_loadings.shape

We created a dataframe to map the terms of topic_0. 

feature_loadings.sort_values('topic_0', ascending=False).head(10)

# Instantiate linear regression model.
logreg = LogisticRegression()



# Fit on Z_train.
logreg.fit(lda_train, y_train)

# Score on training and testing sets.
print(f'Training Score: {round(logreg.score(lda_train, y_train),4)}.')
print(f'Testing Score: {round(logreg.score(lda_test, y_test),4)}.')
