==> code/04c_Semantic Analysis_LR.ipynb <==
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV

pre = pd.read_csv('../assets/combined_pre_classed.csv', index_col=0)
post = pd.read_csv('../assets/combined_post_classed.csv', index_col=0)
df = pd.concat([pre, post], ignore_index=True)
target = df['disaster']

svd_df = pd.read_csv('../assets/svd_df.csv', index_col=0)

X_train, X_test, y_train, y_test = train_test_split(svd_df,
                                                    target,
                                                    stratify=target)

log_reg = LogisticRegression()

params = {
    'penalty'   : ['l1', 'l2'],
    'C'         : [.5, 1, 2, 3, 4, 5]
}

gs = GridSearchCV(log_reg,
                  param_grid=params,
                  cv=3)

gs.fit(X_train, y_train)

gs.score(X_train, y_train)

gs.score(X_test, y_test)

gs.best_params_
